Fix based on review
- Renamed component
- Add set states to useEffect dependency
