--- react/src/lib/components/WebvizViewToggleContainer/index.ts
@@ -1,1 +1,0 @@
-export { WebvizViewToggleContainer } from "./WebvizViewToggleContainer";
