+++ react/src/lib/index.ts
@@ -27,7 +27,7 @@ import { WebvizPluginWrapper } from "./components/WebvizPluginWrapper";
 import { WebvizSettingsDrawer } from "./components/WebvizSettingsDrawer";
 import { WebvizView } from "./components/WebvizView";
 import { WebvizViewElement } from "./components/WebvizViewElement";
-import { WebvizViewToggleContainer } from "./components/WebvizViewToggleContainer";
+import { ViewVisibilityContainer } from "./components/ViewVisibilityContainer";
 import { WebvizSettingsGroup } from "./components/WebvizSettingsGroup";
 import { WebvizPluginLayoutColumn } from "./components/WebvizPluginLayoutColumn/WebvizPluginLayoutColumn";
 import { WebvizPluginLayoutRow } from "./components/WebvizPluginLayoutRow/WebvizPluginLayoutRow";
@@ -44,7 +44,7 @@ export {
     WebvizSettingsDrawer,
     WebvizView,
     WebvizViewElement,
-    WebvizViewToggleContainer,
+    ViewVisibilityContainer,
     WebvizPluginPlaceholder,
     WebvizSettingsGroup,
     WebvizPluginLayoutColumn,
